% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsoluteRiskReduction.R
\name{absoluteRiskReduction}
\alias{absoluteRiskReduction}
\title{Creates a data frame of absolute risk reductions within risk strata}
\usage{
absoluteRiskReduction(dataKM, timePoint)
}
\arguments{
\item{dataKM}{A list with the Kaplan-Meier estimates within risk strata}

\item{timePoint}{The time point based on which the absolute risk reductions will be calculated}
}
\value{
A data frame with the absolute risk reductions along with confidence intervals
}
\description{
Creates a data frame of absolute risk reductions within risk strata
}
