% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmToPlpData.R
\name{cmToPlpData}
\alias{cmToPlpData}
\title{Transforms a cohortMethodData object into a plpData object}
\usage{
cmToPlpData(cohortMethodData)
}
\arguments{
\item{cohortMethodData}{The cohortMethodData object to be transformed}
}
\value{
A plpData object
}
\description{
Transforms a cohortMethodData object, as generated from \code{\link[CohortMethod]{getDbCohortMethodData}}, into a plpData object
appropriate for risk stratified analysis
}
