% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePlp.R
\name{combinePlp}
\alias{combinePlp}
\title{Combine two plpData objects}
\usage{
combinePlp(plpData1File, plpData2File, file)
}
\arguments{
\item{plpData1File}{File location of the first plpDAta object}

\item{plpData2File}{File location of the second plpDAta object}

\item{file}{File location to save the combined plpData pbject}
}
\value{
A combined plpData object and the map matrix
}
\description{
Combine two plpData objects
}
