% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparisonPlot.R
\name{comparisonPlot}
\alias{comparisonPlot}
\title{Generates a comparison plot within risk quantiles}
\usage{
comparisonPlot(dataARR, dataRRR, cases, ylimARR, ylimRRR, ylimCases = c(-1,
  1), legendPositionRisk = c(0.15, 0.15), legendPositionRates = c(0.15,
  0.15), position_dodge.w = 0.2)
}
\arguments{
\item{dataARR}{A list or dataframe same as the one generated by \code{\link[RiskStratifiedEstimation]{absoluteRiskReduction}}}

\item{dataRRR}{A list or dataframe same as the one generated by \code{\link[RiskStratifiedEstimation]{relativeRiskReduction}}}

\item{cases}{The dataframe with the outcome rates within risk strata}

\item{ylimARR}{The limits on the y-axis of the sbsolute risk reduction plot}

\item{ylimRRR}{The limits on the y-axis of the hazard ratio plot}

\item{ylimCases}{The limits on the y-axis of the outcome rates bar plot}

\item{position_dodge.w}{The amount of space between bullets in the ploit in the case of multiple outcomes}

\item{legend.position}{The position of the legend in the plot}
}
\value{
A plot with hazard ratios and absolute risk reductions across risk strata
}
\description{
Generates a plot with relative and absolute risk reductions. It can handle one or multiple outcomes. In the case of multiple outcomes the input must be a named list.
}
