% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonPlotGeneral.R
\name{comparisonPlotGeneral}
\alias{comparisonPlotGeneral}
\title{Creates a general comparison plot}
\usage{
comparisonPlotGeneral(dataARR, dataHR, position_dodge.w = 0.2,
  legend.position = c(0.15, 0.1), ylimHR = c(0, 5), ylimARR = c(-1, 1))
}
\arguments{
\item{dataARR}{A dataframe with columns same as the output of \code{\link[RiskStratifiedEstimation]{absoluteRiskReductionDataFrame}} containing information on all risk models of interest}

\item{dataHR}{A dataframe with columns same as the output of \code{\link[RiskStratifiedEstimation]{hazardRatioDataFrame}} containing information on all risk models of interest}

\item{position_dodge.w}{The position dodge in the graph}

\item{legend.position}{The position of the legend in the graph}

\item{ylimHR}{The limits on the y-axis for the hazard ratio plot}

\item{ylimARR}{The limits on the y-axis for the absolute risk reduction plot}
}
\value{
A combined absolute and relative risk reduction plot
}
\description{
Creates a comparison plot with hazard ratios and absolute risk reductions across risk quantiles
}
