% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonPlotGeneral.R
\name{comparisonPlotGeneral}
\alias{comparisonPlotGeneral}
\title{Generates a comparison plot within risk quantiles}
\usage{
comparisonPlotGeneral(dataARR, dataHR, ylimHR = c(0, 5), ylimARR = c(-1, 1))
}
\arguments{
\item{dataARR}{A dataframe same as the one generated by \code{\link[RiskStratifiedEstimation]{absoluteRiskReduction}}}

\item{dataHR}{A dataframe same as the one generated by \code{\link[RiskStratifiedEstimation]{relativeRiskReduction}}}

\item{ylimHR}{The limits on the y-axis of the hazard ratio plot}

\item{ylimARR}{The limits on the y-axis of the sbsolute risk reduction plot}
}
\value{
A plot with hazard ratios and absolute risk reductions across risk strata
}
\description{
Generates a plot with hazard ratios and absolute risk reductions along with confidence intervals
}
