% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Balance.R
\name{computeCovariateBalance}
\alias{computeCovariateBalance}
\title{Compute covariate balance}
\usage{
computeCovariateBalance(
  population,
  cohorts,
  covariates,
  covariateRef,
  subgroupCovariateId = NULL
)
}
\arguments{
\item{population}{A data frame containing the people that are
remaining after matching and/or trimming.}

\item{cohorts}{The cohorts of a \code{CohortMethodData} object.}

\item{covariates}{The covariates of a \code{CovariateData} object.}

\item{covariateRef}{The covariate reference of a \code{CovariateData}
object.}

\item{subgroupCovariateId}{Optional: a covariate ID of a binary covariate that
indicates a subgroup of interest. Both the before
and after populations will be restricted to this
subgroup before computing covariate balance.}
}
\value{
A tibble with the covariate balance before and after
                            matching/trimming.
}
\description{
Compute covariate balance before and after
                            adjustment with the propensity scores.
}
