% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{computeIncidence}
\alias{computeIncidence}
\title{Compute incidence}
\usage{
computeIncidence(
  population,
  alpha = 0.05,
  power = 0.8,
  twoSided = TRUE,
  modelType = "cox"
)
}
\arguments{
\item{population}{A data frame describing the study population as created using the
\code{\link[CohortMethod]{createStudyPopulation}} function.
This should at least have these columns: personSeqId, treatment,
outcomeCount, timeAtRisk.}

\item{alpha}{Type I error.}

\item{power}{1 - beta, where beta is the type II error.}

\item{twoSided}{Consider a two-sided test?}

\item{modelType}{The type of outcome model that will be used. Possible values
are "logistic", "poisson", or "cox". Currently only "cox"
is supported.}
}
\value{
A tibble with the derived incidence.
}
\description{
Computes the incidence for a specific population
}
