% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{computeIncidenceOverall}
\alias{computeIncidenceOverall}
\title{Compute risk stratified incidence}
\usage{
computeIncidenceOverall(
  path,
  alpha = 0.05,
  power = 0.8,
  twoSided = TRUE,
  modelType = "cox"
)
}
\arguments{
\item{path}{The path to the propensity scores object (list with propensity
scores for each risk stratum).}

\item{alpha}{Type I error.}

\item{power}{1 - beta, where beta is the type II error.}

\item{twoSided}{Consider a two-sided test?}

\item{modelType}{The type of outcome model that will be used. Possible values
are "logistic", "poisson", or "cox". Currently only "cox"
is supported.}
}
\description{
Compute risk stratified incidence for a specific stratification
                 outcome.
}
