% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Balance.R
\name{computeRseeCovariateBalance}
\alias{computeRseeCovariateBalance}
\title{Compute risk stratified covariate balance}
\usage{
computeRseeCovariateBalance(analysisSettings, getDataSettings)
}
\arguments{
\item{analysisSettings}{The analysis settings. Should be created using
\code{\link[RiskstratifiedEstimation]{createAnalysisSettings}}}

\item{getDataSettings}{The \code{getDataSettings} object with the
\code{cohortMethodDataFolder} pointing to the location
where the \code{CohortMethodData} object is stored.}
}
\value{
There is no value returned. The covariate balance results
                        are stored in the proper locations.
}
\description{
Computes covariate balance within strata of predicted
                        risk for all specified analyses
}
