% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{computeRseeIncidence}
\alias{computeRseeIncidence}
\title{Compute risk stratified incidence for all analyses}
\usage{
computeRseeIncidence(analysisSettings)
}
\arguments{
\item{analysisSettings}{An \code{analysisSettings} object created with
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\description{
Computes risk stratified incidence for all the analyses that were
             defined when setting up the study.
}
