% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createAnalysisMapMatrix}
\alias{createAnalysisMapMatrix}
\title{Create map matrix}
\usage{
createAnalysisMapMatrix(id, name, type)
}
\arguments{
\item{id}{The exposure/outcome id (numeric)}

\item{name}{The exposure/outcome name (string)}

\item{type}{Whether it is an "exposure" or "outcome"}
}
\description{
Creates a map matrix for treatments (exposures) and outcomes to be passed to
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.
}
