% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArgFuns.R
\name{createCreateIPWArgs}
\alias{createCreateIPWArgs}
\title{Create a parameter object for the function createIPW
Create an object defining the parameter values.}
\usage{
createCreateIPWArgs(weightsType = "ATE", useStabilizedWeights = TRUE,
  truncationLevels = c(0.01, 0.99))
}
\arguments{
\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for
average treatment effect and 'ATT' for average treatment effect on the
treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity
score.}
}
\description{
Create a parameter object for the function createIPW
Create an object defining the parameter values.
}
