% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createDatabaseSettings}
\alias{createDatabaseSettings}
\title{Create parameter object for database to be reached}
\usage{
createDatabaseSettings(
  cdmVersion = "5",
  cdmDatabaseSchema,
  databaseName,
  cohortDatabaseSchema = cdmDatabaseSchema,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = cdmDatabaseSchema,
  cohortTable,
  outcomeTable,
  exposureTable = "drug_era",
  mergedCohortTable,
  targetCohortId = 1
)
}
\arguments{
\item{cdmVersion}{Define the OMOP CDM version}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this
should specify both the database and the schema, so for example
"cdm_instance.dbo"}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort data used to define the at risk cohort is available. If
cohortTable = DRUG_ERA, \code{cohortDatabaseSchema} is not used by
assumed to be \code{cdmDatabaseSchema}.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the data
used to define the outcome cohorts is available. If cohortTable =
CONDITION_ERA, exposureDatabaseSchema is not used by assumed to be
cdmSchema. Requires read permissions to this database.}

\item{resultsDatabaseSchema}{The name of the database schema with write permissions.}

\item{exposureDatabaseSchema}{Input of function \code{\link[CohortMethod]{getDbCohortMethodData}}:
The name of the database schema that is the location where the
exposure data used to define the exposure cohorts is available.}

\item{cohortTable}{The tablename that contains the at risk cohort. If cohortTable <>
DRUG_ERA, then expectation is cohortTable has format of COHORT
table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{outcomeTable}{The tablename that contains the outcome cohorts. If outcomeTable <>
CONDITION_OCCURRENCE, then expectation is outcomeTable has format of
COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{exposureTable}{The tablename that contains the exposure cohorts. If exposureTable
<> DRUG_ERA, then expectation is exposureTable has format of COHORT
table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{mergedCohortTable}{The name of the table where the merged treatment and comparator
cohorts will be stored.}

\item{targetCohortId}{The cohort definition id of of the merged cohort in the
\code{mergedCohortTable}.}
}
\description{
Create parameter object for database to be reached
}
