% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createGetCovariateSettings}
\alias{createGetCovariateSettings}
\title{Create the parameter object for extracting the covariates}
\usage{

  createGetCovariateSettings(covariateSettingsCm = FeatureExtraction::createCovariateSettings(),
  covariateSettingsPlp = FeatureExtraction::createCovariateSettings())
}
\arguments{
\item{covariateSettingsCm}{The covariate settings object related to the estimation step created from
\code{\link[FeatureExtraction]{createCovariateSettings}}.}

\item{covariateSettingsPlp}{The covariate settings objec related to the prediction step created from
\code{\link[FeatureExtraction]{createCovariateSettings}}.}
}
\description{
Contains the arguments for the extraction of both the covariates related to the prediction step and the estimation
step.
}
