% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArgFuns.R
\name{createGetDataSettings}
\alias{createGetDataSettings}
\title{Create parameter object for extracting the data}
\usage{
createGetDataSettings(getPlpDataSettings = createGetPlpDataArgs(),
  getCmDataSettings = createGetCmDataArgs(), plpDataFolder = NULL,
  cohortMethodDataFolder = NULL)
}
\arguments{
\item{getPlpDataSettings}{Parameter object for the extraction of the \code{plpData} object
created from
\code{\link[RiskStratifiedEstimation]{createGetPlpDataArgs}}.}

\item{getCmDataSettings}{Parameter object for the extraction of the \code{cohortMethodData}
object created from
\code{\link[RiskStratifiedEstimation]{createGetPlpDataArgs}}.}

\item{plpDataFolder}{The directory path where the \code{plpData} object has already been
saved locally.}

\item{cohortMethodDataFolder}{The directory path where the \code{cohortMethodData} object has
already been saved locally.}
}
\description{
Create parameter object for extracting the data
}
