% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArgFuns.R
\name{createGetPlpDataArgs}
\alias{createGetPlpDataArgs}
\title{Create a parameter object for the function getPlpData}
\usage{
createGetPlpDataArgs(studyStartDate = "", studyEndDate = "",
  firstExposureOnly = FALSE, washoutPeriod = 0,
  excludeDrugsFromCovariates = FALSE)
}
\arguments{
\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index date can
appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index date can
appear. Date format is 'yyyymmdd'. Important: the studyend data is also
used to truncate risk windows, meaning no outcomes beyond the study end
date will be considered.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that this is
typically done in the createStudyPopulation function, but can already be
done here for efficiency reasons.}

\item{washoutPeriod}{The mininum required continuous observation time prior to index date for
a person to be included in the at risk cohort. Note that this is
typically done in the createStudyPopulation function,but can already be
done here for efficiency reasons.}

\item{excludeDrugsFromCovariates}{A redundant option}
}
\description{
Create a parameter object for the function getPlpData
}
\details{
Create an object defining the parameter values.
}
