% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIPW.R
\name{createIPW}
\alias{createIPW}
\title{Calcuates inverse probability weights based on the propensity score}
\usage{
createIPW(ps, weightsType = "ATE", truncatedWeights = TRUE, useSW = TRUE,
  truncationQuantiles = c(0.01, 0.99))
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{truncatedWeights}{Should truncated weights be used?}

\item{useSW}{Should stabilized weights be used?}

\item{truncationQuantiles}{Quantiles to truncate the weights}
}
\value{
The ps data frame provided as input along with a weights column
}
\description{
Calcuates inverse probability weights based on the propensity score
}
