% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIPW.R
\name{createIPW}
\alias{createIPW}
\title{Creates Inverse Probability Weights}
\usage{
createIPW(ps, weightsType = "ATE", useStabilizedWeights = TRUE,
  extremeWeights = NULL, truncationLevels = c(0.01, 0.99),
  cvLikeRepetitions, stepTruncationLevels)
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{extremeWeights}{The way to assess extreme weights. Possible options are 'unadjusted, 'cvLikeTruncation', 'crumpTrimming', 'fixedTruncation'}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score. If extremeWeights is 'fixedTruncation' then the weights will be truncated at the levels defined here. If extremeWeights is 'cvLikeTruncation' then the data adaptive procedure will only assess truncation up to the levels defined here}

\item{cvLikeRepetitions}{The number of times to repeat the 2-fold cross-validations}

\item{stepTruncationLevels}{The steps for the grid of possible truncation levels}
}
\value{
The ps data frame provided as input along with a weights column
}
\description{
Calcuates inverse probability weights based on the propensity score
}
