% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{createOverallResults}
\alias{createOverallResults}
\title{Combines the overall results}
\usage{
createOverallResults(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}
}
\value{
Stores the overall results along with the required data to lauch the shiny
                                  application in the `shiny` directory
}
\description{
Combines the overall results
}
