% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunCmAnalysesArgs}
\alias{createRunCmAnalysesArgs}
\title{Create a parameter object for defining the estimation analyses.}
\usage{
createRunCmAnalysesArgs(
  psMethod = "stratifyByPs",
  label = NULL,
  effectEstimationSettings = createStratifyByPsArgs(),
  psSettings = createCreatePsArgs(),
  timePoint = 365
)
}
\arguments{
\item{psMethod}{How should the propensity scores be used? Can
be one of "inversePtWeighted", "stratifyByPs"
 or "matchOnPs".}

\item{label}{A single-word description of the analysis.}

\item{effectEstimationSettings}{Parameter object providing further settings
for the implementation of selected \code{psMethod}
to the estimation process. Can be created using
one of
\code{\link[RiskStratifiedEstimation]{createCreateIPWArgs}},
when \code{inversePtWeighted} is selected,
\code{\link[RiskStratifiedEstimation]{createStratifyByPsArgs}}
when \code{stratifyByPs} is selected or
\code{\link[RiskStratifiedEstimation]{createMatchOnPsArgs}}
when \code{matchOnPs} is selected.}

\item{psSettings}{Parameter object for
\code{\link[CohortMethod]{createPs}}}

\item{timePoint}{The time point after cohort start that absolute differences
should be estimated.}
}
\value{
A parameter object for running the the estimation step.
}
\description{
Create a parameter object for defining the estimation analyses.
}
