% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunCmAnalysis}
\alias{createRunCmAnalysis}
\title{Create a parameter object for defining the estimation analyses.}
\usage{
createRunCmAnalysis(
  label,
  description = "",
  stratificationOutcome,
  estimationOutcomes,
  riskStratificationMethod,
  psAdjustmentMethod,
  timePoint
)
}
\arguments{
\item{label}{Label of the analysis (no spaces allowed).}

\item{description}{Description of the analysis.}

\item{stratificationOutcome}{The outcome ID on which risk stratification
will be carried out. Should be one of the
stratification outcomes defined in the
`analysisSettings`.}

\item{estimationOutcomes}{Outcomes to estimate risk-stratified treatment
effects.}

\item{riskStratificationMethod}{Risk stratification method}

\item{psAdjustmentMethod}{Propensity score adjustment method}

\item{timePoint}{The time point after cohort start that absolute differences
should be estimated.}
}
\value{
A parameter object for running the the estimation step.
}
\description{
Create a parameter object for defining the estimation analyses.
}
