% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunCmSettingsArgs}
\alias{createRunCmSettingsArgs}
\title{Create a parameter object for running the estimation step Create a parameter
object for running the estimation step. This function is used to create part
of the input of \code{\link[RiskStratifiedEstimation]{createRunSettings}}.}
\usage{
createRunCmSettingsArgs(
  psMethod = "stratifyByPs",
  label = NULL,
  effectEstimationSettings = createStratifyByPsArgs(),
  psSettings = createCreatePsArgs(),
  createPsThreads = 1,
  fitOutcomeModelsThreads = 1,
  timePoint = 365,
  riskStrata = 4
)
}
\arguments{
\item{psMethod}{How should the propensity scores be used? Can
be one of "inversePtWeighted", "stratifyByPs"
 or "matchOnPs".}

\item{label}{A single-word description of the analysis.}

\item{effectEstimationSettings}{Parameter object providing further settings
for the implementation of selected \code{psMethod}
to the estimation process. Can be created using
one of
\code{\link[RiskStratifiedEstimation]{createCreateIPWArgs}},
when \code{inversePtWeighted} is selected,
\code{\link[RiskStratifiedEstimation]{createStratifyByPsArgs}}
when \code{stratifyByPs} is selected or
\code{\link[RiskStratifiedEstimation]{createMatchOnPsArgs}}
when \code{matchOnPs} is selected.}

\item{psSettings}{Parameter object for
\code{\link[CohortMethod]{createPs}}}

\item{createPsThreads}{The number of parallel threads for the
estimation of the propensity scores. Default is 1.}

\item{fitOutcomeModelsThreads}{The number of parallel threads for the estimation of the
outcome models.}

\item{timePoint}{The time point after cohort start that absolute differences
should be estimated.}

\item{riskStrata}{The number of risk strata. Default is 4.}
}
\value{
A parameter object for running the the estimation step.
}
\description{
Create a parameter object for running the estimation step Create a parameter
object for running the estimation step. This function is used to create part
of the input of \code{\link[RiskStratifiedEstimation]{createRunSettings}}.
}
