% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunExistingPlpSettingsArgs}
\alias{createRunExistingPlpSettingsArgs}
\title{Create prediction settings for an existing model}
\usage{
createRunExistingPlpSettingsArgs(
  outcomeId,
  plpResultDirectory,
  predictionSettings = list()
)
}
\arguments{
\item{outcomeId}{The outcome ID of the existing model.}

\item{plpResultDirectory}{The directory where the existing model is store.d}

\item{predictionSettings}{A list with prediction settings.}
}
\description{
Create prediction settings for an existing model
}
