% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunPlpAnalysesArgs}
\alias{createRunPlpAnalysesArgs}
\title{Create a parameter object for the function runPlp}
\usage{
createRunPlpAnalysesArgs(
  outcomeId = NULL,
  splitSettings = PatientLevelPrediction::createDefaultSplitSetting(),
  sampleSettings = PatientLevelPrediction::createSampleSettings(),
 
    featureEngineeringSettings = PatientLevelPrediction::createFeatureEngineeringSettings(),
  preprocessSettings = PatientLevelPrediction::createPreprocessSettings(),
  modelSettings = PatientLevelPrediction::setLassoLogisticRegression(),
  logSettings = PatientLevelPrediction::createLogSettings(),
  executeSettings = PatientLevelPrediction::createDefaultExecuteSettings(),
  matchingSettings = CohortMethod::createMatchOnPsArgs(),
  timepoint = NULL
)
}
\arguments{
\item{outcomeId}{The outcomeId for the prediction.}

\item{matchingSettings}{The settings for the construction of the population on which the
prediction model will be developed.}

\item{timepoint}{The timepoint to predict risk (survival models only)}

\item{spliteSettings}{The split settings.}
}
\description{
Create a parameter object for the function runPlp
}
\details{
Create an object defining the parameter values.
}
