% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunPlpSettingsArgs}
\alias{createRunPlpSettingsArgs}
\title{Create run settings for prediction}
\usage{
createRunPlpSettingsArgs(
  analyses = list(),
  defaultSettings = createRunPlpAnalysesArgs()
)
}
\arguments{
\item{analyses}{A list with settings for the prediction of each of
the risk stratification outcomes.}

\item{defaultSettings}{An object of type \code{runPlpAnalysesArgs} used as
the default settings for outcome risk prediction. Should
be created with
\code{\link[RiskStratifiedEstimation]{createRunPlpAnalysesArgs}}}
}
\value{
An object of type \code{runPlpSettingsArgs} to be passed
                        to \code{\link[RiskStratifiedEstimation]{createRunSettings}}
}
\description{
Create run settings for prediction
}
