% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArgFuns.R
\name{createRunSettings}
\alias{createRunSettings}
\title{Create a settings object for running the analyses}
\usage{
createRunSettings(runPlpArgs = createRunPlpArgs(modelSettings =
  PatientLevelPrediction::setLassoLogisticRegression()),
  runCmArgs = createRunCmArgs())
}
\arguments{
\item{runPlpArgs}{A parameterer object of type \code{runPlpArgs} defined using the function
\code{\link[RiskStratifiedEstimation]{createRunPlpArgs}}.}

\item{runCmArgs}{A parameter object of type \dode{runCmArgs} defined using the function
\code{\link[RiskStratifiedEstimation]{createRunCmArgs}}}
}
\value{
An R object of type \code{runSettings}
}
\description{
Create the settings for running the analyses. The input consists of two parts: 1) the settings for running the
prediction algorithms and 2) the settings for estimating treatment effects within strata of predicted risk.
}
