% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createServer}
\alias{createServer}
\title{Creates the server function of the shiny application}
\usage{
createServer(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\value{
An R file containing the sever function
}
\description{
Creates the server function of the shiny application
}
