% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createUI}
\alias{createUI}
\title{Creates the ui function of the shiny application}
\usage{
createUI(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\value{
An R file containing the ui function
}
\description{
Creates the ui function of the shiny application
}
