% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{estimateTreatmentEffect}
\alias{estimateTreatmentEffect}
\title{Estimate treatment effects within risk strata}
\usage{
estimateTreatmentEffect(ps, runCmSettings)
}
\arguments{
\item{ps}{A list of objects created from \code{\link[CohortMethod]{createPs}} estimated within risk
strata.}

\item{runCmSettings}{The settings for running a \code{CohortMethod} analysis.}
}
\value{
A list containing :
                        \itemize{
                          \item{relativeRiskReduction}{Hazard ratios along with confidence intervals within risk
                            strata}
                          \item{absoluteRiskReduction}{Absolute risk differences along with confidence intervals
                            within risk strata}
                          \item{cases}{Observed outcome proportions within risk strata}
                          \item{models}{The models used to estimate relative risk reduction within risk strata}
                        }
}
\description{
Estimates treatment effects within risk strata based on the length of list \code{ps}.
}
