% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateExistingModelCombine.R
\name{evaluateExistingModelCombine}
\alias{evaluateExistingModelCombine}
\title{Evaluate an existing model}
\usage{
evaluateExistingModelCombine(plpData, outcomeId, riskWindowStart = 1,
  riskWindowEnd = 365, requireTimeAtRisk = T, minTimeAtRisk = 364,
  includeAllOutcomes = T)
}
\arguments{
\item{plpData}{a plpData object}

\item{outcomeId}{the outcome cohort id}

\item{riskWindowStart}{Parameter to be passed to \code{\link[PatientLevelPrediction]{createStudyPopulation}}.The start of the risk window (in days) relative to the index date (+ days of exposure if the addExposureDaysToStart parameter is specified)}

\item{riskWindowEnd}{Parameter to be passed to \code{\link[PatientLevelPrediction]{createStudyPopulation}}. The end of the risk window (in days) relative to the index data (+ days of exposure if the addExposureDaysToEnd parameter is specified)}

\item{requireTimeAtRisk}{Parameter to be passed to \code{\link[PatientLevelPrediction]{createStudyPopulation}}. Should subject without time at risk be removed?}

\item{minTimeAtRisk}{Parameter to be passed to \code{\link[PatientLevelPrediction]{createStudyPopulation}}. The minimum number of days at risk required to be included}

\item{includeAllOutcomes}{Parameter to be passed to \code{\link[PatientLevelPrediction]{createStudyPopulation}}. (binary) indicating whether to include people with outcomes who are not observed for the whole at risk period}
}
\value{
A list containing the performance values
}
\description{
Evaluates an existing model within the setting of a risk stratified analysis
}
