% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall.R
\name{fitOutcomeModelsOverall}
\alias{fitOutcomeModelsOverall}
\title{Fit overall outcome model}
\usage{
fitOutcomeModelsOverall(
  outcomeId,
  analysisSettings,
  getDataSettings,
  runCmSettings
)
}
\arguments{
\item{outcomeId}{The outcome of interest for which the esitmation is performed. That is the outcome for
which risk stratification is performed.}

\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{runCmSettings}{A parameter object of type \code{runCmSettingsArgs} defined using the function
\code{\link[RiskStratifiedEstimation]{createRunCmSettingsArgs}}}
}
\value{
\code{NULL}. The results are all saved.
}
\description{
Fits outcome models within risk strata, estimating relative and absolute differences. Designed to be performed within
a parellelized analysis.
}
