% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{fitPsModel}
\alias{fitPsModel}
\title{Calculate propensity scores for a specific outcome}
\usage{
fitPsModel(cohortMethodDataFolder, plpDataFolder, outcomeId,
  populationCmSettings, populationPlpSettings, saveDirectory, riskStrata,
  analysisId, analysisPath)
}
\arguments{
\item{cohortMethodDataFolder}{The directory where the \code{cohortMethodData} object is stored.}

\item{plpDataFolder}{The directory where the \code{plpData} object is stored.}

\item{outcomeId}{The outcome of interest for which the risk stratification is performed.}

\item{populationCmSettings}{A parameter object for the function \code{\link[CohortMethod]{createStudyPopulation}}.
Can be generated from function \code{createStudyPopulationCmSettings}.}

\item{populationPlpSettings}{A parameter object for the function \code{\link[PatientLevelPrediction]{createStudyPopulation}}.
Can be generated from unction \code{\link[PatientLevelPrediction]{createStudyPopulationSettings}}.}

\item{saveDirectory}{The direcotry of the RSEE analysis where the prediction models are located.}

\item{riskStrata}{The considered number of risk strata.}

\item{analysisId}{The analysis ID of the prediction model used to stratify the population.}

\item{analysisPath}{The directory where the propensity scores will be stored.}
}
\value{
\code{NULL}. The results are all saved.
}
\description{
Fits a large-scale regularized regression model to estimate propensity scores within predicted risk strata. Designed
to be applied in a parallelized analysis.
}
