% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsModels.R
\name{fitPsModelSwitch}
\alias{fitPsModelSwitch}
\title{Fits switched propensity score models}
\usage{
fitPsModelSwitch(
  analysis,
  predictOutcome,
  compareOutcome,
  initialPopulation,
  analysisSettings,
  getDataSettings,
  populationSettings,
  runSettings
)
}
\arguments{
\item{predictOutcome}{The outcome of the prediction step}

\item{compareOutcome}{The outcome of interest for the estimation step}

\item{initialPopulation}{The initial population to be used for the
estimation of the switched-outcome propensity
scores.}

\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{populationSettings}{An R object of type \code{populationSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createPopulationSettings}}.}

\item{runSettings}{An R object of type \code{runSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createRunSettings}}.}
}
\description{
Fits propensity score models where the population is stratified based on prediction model for a certain
outcome of interest and estimation is focused on a different outcome
}
