% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateResults.R
\name{generateResults}
\alias{generateResults}
\title{Generates results within risk strata}
\usage{
generateResults(ps, timePoint)
}
\arguments{
\item{ps}{A list of the estimated propensity scores within risk strata with a column containing the inverse probability weights.
The elements can be either dataframes or \code{ffdf} objects.}

\item{timePoint}{The time point of interest for the calculation of the absolute risk reduction.}
}
\value{
A list containing the weighted Kaplan-Meier estimates, the absolute and relative risk reductions along with the observed case-counts
        wtihin risk strata.
}
\description{
Generates results within risk strata. More specifically, it calculates the weighted Kaplan-Meier estimates, the relative and absolute risk
differences within risk strata.
}
