% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardRatioDataFrame.R
\name{hazardRatioDataFrame}
\alias{hazardRatioDataFrame}
\title{Creates a data frame with hazard ratios within risk strata}
\usage{
hazardRatioDataFrame(psList, weightsType = "ATE", useSW = TRUE,
  truncatedWeights = TRUE, truncationQuantiles = c(0.01, 0.99))
}
\arguments{
\item{psList}{A list with the propensity data frames as generated from \code{\link[CohortMethod]{createPs}}}

\item{weightsType}{The type of weights to be used. Possible values are 'ATE' or 'ATT'}

\item{useSW}{Whether to use stabilized weights.}

\item{truncatedWeights}{Whether to truncate the weights}

\item{truncationQuantiles}{The quantiles used to truncate the weights}
}
\value{
A data frame with hazard ratios along with confidence intervals
}
\description{
Creates a data frame with hazard ratios within risk strata
}
