% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{outcomeModelWeighted}
\alias{outcomeModelWeighted}
\title{Fits a weighted cox regression model}
\usage{
outcomeModelWeighted(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, truncationLevels)
}
\arguments{
\item{ps}{A dataframe wiht propensity scores as generated from \code{\link[CohortMethod]{createPs}}.}

\item{calculateWeights}{Should weights be calculated?}

\item{weightsType}{The type of weights for the balancing of covariates. Should be either 'ATE' or 'ATT'}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score.}
}
\value{
A weighted cox regression model.
}
\description{
Fits a weighted cox regression model using an inverse probability of treatment weighting approach
}
