% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalibration.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Get the Calibration Plot of a Prediction Model}
\usage{
plotCalibration(prediction, nQuantiles = 10, lims = c(0, 1))
}
\arguments{
\item{prediction}{A prediction dataframe as returned by \code{\link[PatientLevelPrediction]{runPlp}}. The predicted probabilities should be under the name "value"}

\item{nQuantiles}{The number of quantiles of predicted risk to split the data}

\item{lims}{The limits on the x and y-axis. Cannot be lower than 0 or higher than 1}
}
\value{
The calibration plot and the calibration summary
}
\description{
Get the calibration plot and the calibration summary of a prediction model
}
