% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotRSEE}
\alias{plotRSEE}
\title{Plot the results of a risk stratified analysis}
\usage{
plotRSEE(rseeResult, mapOutcomes = NULL, mapTreatments = NULL)
}
\arguments{
\item{rseeResult}{The overall result of a risk stratified analysis.}

\item{mapOutcomes}{A dataframe with the outcome labels. It should have 2 columns called "outcome" and "labelOutcome".
The former should contain character values of the form "outcome_x" and the latter should contain
the outcome label. If set to \code{NULL}, the outcome defintion id's appear in the graphs.}

\item{mapTreatments}{A dataframe with the treatment labels. It should have 2 columns named "cohort" and "labelTreatments".
The former should contain the values "treatment" and "comparator" and the latter should contain
the treatment labels. If set to \code{NULL}, the labels treatment and comaprator appear in
the graph,}
}
\value{
A 3-level graph. In the first level observed outcome rates are presented for all treatment-outcome
                                     combinations across risk strata. In the second level, hazard ratios across risk strata are given.
                                     In the final level, absolute risk difference across risk strata are presented.
}
\description{
Plots the overall results of a risk stratified analysis.
}
