% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeightedKM.R
\name{plotWeightedKM}
\alias{plotWeightedKM}
\title{Plots the weighted Kaplan-Meier estimate}
\usage{
plotWeightedKM(dataKM, xlim, ylim, legend.position = c(0.15, 0.1),
  ci = TRUE, title = NULL)
}
\arguments{
\item{dataKM}{A dataframe containing the weigthed Kaplan-Meier estimates as computed by \code{\link[RiskStratifiedEstimation]{weightedKM}}}

\item{xlim}{Limits on x-axis}

\item{ylim}{Limits on y-axis}

\item{legend.position}{The position of the legend on the graph}

\item{ci}{Should confidence intervals be displayed?}

\item{title}{The title on the graph}
}
\value{
A weighted Kaplan-Meier plot
}
\description{
Plots the weighted Kaplan-Meier estimate
}
