% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionSummary.R
\name{predictionSummary}
\alias{predictionSummary}
\title{Prediction Summary}
\usage{
predictionSummary(predictionData, response, predictors, nFolds = 4,
  nQuantiles = 10, lims = c(0, 1))
}
\arguments{
\item{predictionData}{The data frame from which the prediction will be calculated}

\item{predictors}{The vector with the names of the predictor variables}

\item{nFolds}{The number of the folds to be used in cross validation}

\item{nQuantiles}{The number of risk quantiles to be considered}

\item{lims}{The axes limits on the calibration plot}

\item{repsonse}{The name of the response variable}
}
\value{
The AUC, calibration summary and the calibration plot
}
\description{
Calulates the AUC and the calibration results based on cross validation
}
