% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{prepareMultipleRseeViewer}
\alias{prepareMultipleRseeViewer}
\title{Prepare shiny viewer}
\usage{
prepareMultipleRseeViewer(analysisSettingsList, saveDirectory)
}
\arguments{
\item{analysisSettingsList}{A list of all the targeted \code{analysisSettings}.}

\item{saveDirectory}{The directory where the results will be written.}
}
\description{
Combines results from multiple analyses for the shiny application.
}
