% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskReduction.R
\name{relativeRiskReduction}
\alias{relativeRiskReduction}
\title{Creates a data frame with hazard ratios within risk strata}
\usage{
relativeRiskReduction(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, truncationLevels)
}
\arguments{
\item{ps}{A list with the propensity data frames as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score.}
}
\value{
A data frame with hazard ratios along with confidence intervals
}
\description{
Creates a data frame with hazard ratios within risk strata
}
