% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{rseeViewer}
\alias{rseeViewer}
\title{Risk stratified effect estimation viewer}
\usage{
rseeViewer(analysisPath)
}
\arguments{
\item{analysisPath}{The shiny directory where the results are stored.}
}
\description{
Launches a the shiny application for the exploration of results
}
