% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{runRiskStratifiedEstimation1}
\alias{runRiskStratifiedEstimation1}
\title{Runs a risk stratified analysis}
\usage{
runRiskStratifiedEstimation1(connectionDetails, analysisSettings,
  databaseSettings, getDataSettings, covariateSettings, populationSettings,
  runSettings)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function
\code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{analysisSettings}{An R object of type \code(analysisSettings) created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}

\item{databaseSettings}{An R object of type \code{databaseSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createDatabaseSettings}}.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{covariateSettings}{An R object of type \code{covariateSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createCovariateSettings}}.}

\item{populationSettings}{An R object of type \code{populationSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createPopulationSettings}}.}

\item{runSettings}{An R object of type \code{runSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createRunSettings}}.}
}
\value{

}
\description{
Runs a risk stratified analysis in two stages. It first runs a prediction algorithm using
\code{PatientLevelPrediction} to derive baseline patient risks and then derives estimates within risk strata using
\code{CohortMethod} package.
}
