% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{runShiny}
\alias{runShiny}
\title{Launches the shiny application}
\usage{
runShiny(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\value{
An R file containing the sever function
}
\description{
Launches the shiny application that enables the exploration of the risk stratified analysis
}
