% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{singlePlotRSEE}
\alias{singlePlotRSEE}
\title{Plot a single outcome}
\usage{
singlePlotRSEE(outcomeId, mapMatrix, overallResult = NULL,
  mapTreatments = NULL)
}
\arguments{
\item{outcomeId}{The outcome Id of interest}

\item{mapMatrix}{The matrix that maps patients to their risk quantiles}

\item{overallResult}{The hazard ratio from an analysis on the overall study population}

\item{mapTreatments}{A dataframe with the treatment labels. It should have 2 columns named "cohort" and "labelTreatments".
The former should contain the values "treatment" and "comparator" and the latter should contain
the treatment labels. If set to \code{NULL}, the labels treatment and comaprator appear in
the graph.}
}
\value{
A 3-level graph for a single outcome. In the first level observed outcome rates are presented
                                 for all treatment-outcome combinations across risk strata. In the second level, hazard ratios
                                 across risk strata are given. In the final level, absolute risk difference across risk strata are presented.
}
\description{
Plots the result of a risk stratified analysis for a single outcome of interest
}
