% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{stratifiedKaplanMeier}
\alias{stratifiedKaplanMeier}
\title{Calculate stratified Kaplan-Meier estimates}
\usage{
stratifiedKaplanMeier(population, timePoint)
}
\arguments{
\item{population}{The population of interest stratified using the  \code{\link[CohortMethod]{stratifyByPs}}}

\item{timePoint}{The point in time for which the absolute risk difference is required}
}
\value{
A vector of the absolute risk difference along with the lowest and highest limits of the
                                 the 95 percent confidence interval
}
\description{
Calculate stratified Kaplan-Meier estimates
}
