% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{weightedKM}
\alias{weightedKM}
\title{Calculates and plots weighted Kaplan-Meier estimates}
\usage{
weightedKM(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, extremeWeights,
  fixedTruncationLevels = c(0.01, 0.99), truncationLevels = 0.1,
  cvLikeRepetitions, stepTruncationLevels)
}
\arguments{
\item{ps}{A data frame including the propensity scores as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{extremeWeights}{The way to assess extreme weights. Possible options are 'unadjusted, 'cvLikeTruncation', 'crumpTrimming', 'fixedTruncaiton'}

\item{fixedTruncationLevels}{The levels for fixed truncation weighting}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score. Only symmetric truncation is available. E.g. truncationLevels =.01 will assess truncation up to the .99th percentile of ps}

\item{cvLikeRepetitions}{The number of times to repeat the 2-fold cross-validations}

\item{stepTruncationLevels}{The steps for the grid of possible truncation levels}
}
\value{
A data frame with the Kaplan-Meier estimates
}
\description{
Calculates the weighted Kaplan-Meier estimates based on: Xie J, Liu C. Adjusted Kaplan-Meier estimator and log-rank test with inverse probability of treatment weighting for survival data. Statistics in Medicine 2005; 2:3089–3110.
}
