% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{absoluteRiskReduction}
\alias{absoluteRiskReduction}
\title{Absolute risk reduction}
\usage{
absoluteRiskReduction(population, timePoint, psMethod)
}
\arguments{
\item{population}{The study population generated by \code{\link[CohortMethod]{matchOnPs}} when using
propensity score matching or by \code{\link[CohortMethod]{stratifyByPs}} when stratifying
on the propensity score. In case of inverse probability of treatment weighting approach,
it is a datframe with a \code{weights} column.}

\item{timePoint}{The time at which the absolute risk difference is estimated}

\item{psMethod}{Can be one of "matchOnPs", "stratifyByPs" or "inversePtWeighted".}
}
\value{
A dataframe with the absolute risk-stratum specific absolute risk difference estimates,
                          along with 95 percent confidence interval.
}
\description{
Calculates absolute risk reduction based on the Kaplan-Meier estimates within risk strata
}
