% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createGlobal}
\alias{createGlobal}
\title{Creates the global function of the shiny application}
\usage{
createGlobal(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\value{
An R file containing the global function
}
\description{
Creates the global function that is executed at the launch of the application
}
