% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsModels.R
\name{createIPW}
\alias{createIPW}
\title{Creates Inverse Probability Weights}
\usage{
createIPW(
  ps,
  weightsType = "ATE",
  useStabilizedWeights = TRUE,
  truncationLevels = c(0.01, 0.99)
)
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment
effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score.}
}
\value{
The ps data frame provided as input along with a weights column
}
\description{
Calcuates inverse probability weights based on the propensity score
}
