% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createOverallResults}
\alias{createOverallResults}
\title{Merges the results of a risk stratified analysis}
\usage{
createOverallResults(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}}
}
\value{
An R file containing the sever function
}
\description{
Merges the results of a risk stratified analysis
}
