% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createPopulationSettings}
\alias{createPopulationSettings}
\title{Create parameter object for defining the analysis populations}
\usage{

  createPopulationSettings(populationPlpSettings = createPopulationPlpSettingsArgs(),
  populationCmSettings = createPopulationCmSettingsArgs())
}
\arguments{
\item{populationPlpSettings}{Parameter object for the definition of the \code{populationPlp} object created from
\code{\link[RiskStratifiedEstimation]{createPopulationPlpSettingsArgs}}.}

\item{populationCmSettings}{Parameter object for the definition of the \code{populationCm} object created from
\code{\link[RiskStratifiedEstimation]{createPopulationCmSettingsArgs}}.}
}
\description{
Contains the settings for defining both the \code{populationPlp} and the \code{populationCm} objects.
}
