% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFuns.R
\name{createRunCmSettingsArgs}
\alias{createRunCmSettingsArgs}
\title{Create a parameter object for running the estimation step}
\usage{
createRunCmSettingsArgs(
  analyses,
  psSettings = createCreatePsArgs(),
  createPsThreads = 1,
  fitOutcomeModelsThreads = 1,
  balanceThreads = 1,
  negativeControlThreads = 1
)
}
\arguments{
\item{analyses}{A list of the analyses to run. Each element of
the list can be creaeted using
\code{\link[RiskStratifiedEstimation]{createRunCmAnalysesArgs}}
\code{\link[CohortMethod]{createPs}}}

\item{psSettings}{The settings for estimating the propensity scores}

\item{createPsThreads}{The number of parallel threads for the
estimation of the propensity scores. Default is 1.}

\item{fitOutcomeModelsThreads}{The number of parallel threads for the estimation of the
outcome models.}

\item{balanceThreads}{The number of parallel threads for the estimation
of covariate balance}

\item{negativeControlThreads}{The number of parallel threads for the negative
control analyses}
}
\value{
A parameter object for running the the estimation step.
}
\description{
Create a parameter object for running the estimation step. This function is
used to create the computational part of the input of
\code{\link[RiskStratifiedEstimation]{createRunSettings}}.
}
