% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArgFuns.R
\name{createRunSettings}
\alias{createRunSettings}
\title{Create a settings object for running the analyses
Create the settings for running the analyses. The input consists of two parts: 1) the settings for
running the prediction algorithms and 2) the settings for estimating treatment effects within
strata of predicted risk.}
\usage{

  createRunSettings(runPlpSettings = createRunPlpSettingsArgs(modelSettings
  = PatientLevelPrediction::setLassoLogisticRegression()),
  runCmSettings = createRunCmSettingsArgs())
}
\arguments{
\item{runPlpSettings}{A parameterer object of type \code{runPlpSettingsArgs} defined using the
function \code{\link[RiskStratifiedEstimation]{createRunPlpSettingsArgs}}.}

\item{runCmSettings}{A parameter object of type \code{runCmSettingsArgs} defined using the
function \code{\link[RiskStratifiedEstimation]{createRunCmSettingsArgs}}}
}
\value{
An R object of type \code{runSettings}
}
\description{
Create a settings object for running the analyses
Create the settings for running the analyses. The input consists of two parts: 1) the settings for
running the prediction algorithms and 2) the settings for estimating treatment effects within
strata of predicted risk.
}
