% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvLikeTruncation.R
\name{cvLikeTruncation}
\alias{cvLikeTruncation}
\title{Calculates the cv-like truncation level}
\usage{
cvLikeTruncation(ps, truncationLevels, stepTruncationLevels,
  cvLikeRepetitions = 50)
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score. Only symmetric truncation is available. E.g. truncationLevels =.01 will assess truncation up to the 99th percentile of ps}

\item{stepTruncationLevels}{The steps for the grid of possible truncation levels}

\item{cvLikeRepetitions}{The number of times to repeat the 2-fold cross-validations}
}
\value{
The truncation quantile
}
\description{
Calculates the cv-like truncation level using 2-fold cross-validation for a specified number of repetitions
}
