% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{fitOutcomeModels}
\alias{fitOutcomeModels}
\title{Fit outcome models}
\usage{
fitOutcomeModels(outcomeId, getDataSettings, pathToPs, runSettings)
}
\arguments{
\item{outcomeId}{The outcome of interest for which the esitmation is performed. That is the outcome
for which risk stratification is performed.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{pathToPs}{The path to the \code{RSEE} analysis results.}

\item{runSettings}{An R object of type \code{runSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createRunSettings}}.}
}
\value{
\code{NULL}. The results are all saved.
}
\description{
Fits outcome models within risk strata, estimating relative and absolute differences. Designed to be performed within
a parellelized analysis.
}
