% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsModels.R
\name{fitPsModel}
\alias{fitPsModel}
\title{Calculate propensity scores for a specific outcome}
\usage{
fitPsModel(
  outcomeId,
  initialPopulation,
  runSettings,
  getDataSettings,
  populationSettings,
  analysisSettings
)
}
\arguments{
\item{outcomeId}{The outcome of interest for which the esitmation is performed. That is the outcome for
which risk stratification is performed.}

\item{initialPopulation}{The initial population on the subset of which
the propensity scores will be estimated.}

\item{runSettings}{An R object of type \code{runSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createRunSettings}}.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{populationSettings}{An R object of type \code{populationSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createPopulationSettings}}.}

\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}
}
\value{
\code{NULL}. The results are all saved.
}
\description{
Fits a large-scale regularized regression model to estimate propensity scores within predicted risk strata. Designed
to be applied in a parallelized analysis.
}
