% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectEstimation.R
\name{fitPsModelOverall}
\alias{fitPsModelOverall}
\title{Calculate propensity scores for a specific outcome}
\usage{
fitPsModelOverall(
  outcomeId,
  getDataSettings,
  populationSettings,
  analysisSettings,
  runCmSettings
)
}
\arguments{
\item{outcomeId}{The outcome of interest for which the risk stratification is performed.}

\item{getDataSettings}{An R object of type \code{getDataSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createGetDataSettings}}.}

\item{populationSettings}{An R object of type \code{covariateSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createPopulationSettings}}}

\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}

\item{runCmSettings}{A parameter object of type \code{runCmSettingsArgs} defined using the function
\code{\link[RiskStratifiedEstimation]{createRunCmSettingsArgs}}}
}
\value{
\code{NULL}. The results are all saved.
}
\description{
Fits a large-scale regularized regression model to estimate propensity scores within predicted risk strata. Designed
to be applied in a parallelized analysis.
}
