% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRSEEResult.R
\name{loadRSEEResult}
\alias{loadRSEEResult}
\title{Loads the result from \code{\link{runRiskStratifiedEstimation}}}
\usage{
loadRSEEResult(location)
}
\arguments{
\item{location}{The location wher the result of \code{\link{runRiskStratifiedEstimation}} was saved. By default it is the folder called RSEE along with the date and time of the start}
}
\value{
An object with all the output created from runRiskStratifiedEstimation
}
\description{
Loads the result from \code{\link{runRiskStratifiedEstimation}}
}
