% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCovariateBalance.R
\name{plotCovariateBalance}
\alias{plotCovariateBalance}
\title{Plots the covariate balance before and after balancing}
\usage{
plotCovariateBalance(ps, cohortMethodData, calculateWeights = TRUE,
  weightsType = "ATE", useStabilizedWeights = TRUE,
  truncatedWeights = TRUE, truncationQuantiles = c(0.01, 0.99))
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{cohortMethodData}{A cohortMethodData object}

\item{calculateWeights}{Should the weights be calculated?}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncatedWeights}{Should truncated weights be used? If FALSE, then stabilized weights are used instead}

\item{truncationQuantiles}{The quantiles to perform weight truncation}
}
\value{
The covariate balance plot
}
\description{
Plots covariate before and after weighting using the inverse of the propensity score
}
