% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCovariateBalance.R
\name{plotCovariateBalance}
\alias{plotCovariateBalance}
\title{Plots the covariate balance before and after balancing}
\usage{
plotCovariateBalance(ps, cohortMethodData, calculateWeights = TRUE,
  weightsType = "ATE", useStabilizedWeights = TRUE, extremeWeights,
  fixedTruncationLevels = c(0.01, 0.99), truncationLevels = 0.1,
  cvLikeRepetitions, stepTruncationLevels, showNotBalancedCovariateIds = TRUE)
}
\arguments{
\item{ps}{A propensity score data frame as created from \code{\link[CohortMethod]{createPs}}}

\item{cohortMethodData}{A cohortMethodData object}

\item{calculateWeights}{Should the weights be calculated?}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{extremeWeights}{The way to assess extreme weights. Possible options are 'unadjusted, 'cvLikeTruncation', 'crumpTrimming', 'fixedTruncaiton'}

\item{fixedTruncationLevels}{The levels for fixed truncation weighting}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score. Only symmetric truncation is available. E.g. truncationLevels =.01 will assess truncation up to the .99th percentile of ps}

\item{cvLikeRepetitions}{The number of times to repeat the 2-fold cross-validations}

\item{stepTruncationLevels}{The steps for the grid of possible truncation levels}

\item{showNotBalancedCovariateIds}{Show covariate ids that were not balanced after weighting?}
}
\value{
The covariate balance plot
}
\description{
Plots covariate before and after weighting using the inverse of the propensity score
}
