% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeightedKM.R
\name{plotWeightedKM}
\alias{plotWeightedKM}
\title{Plots the weighted Kaplan-Meier estimate}
\usage{
plotWeightedKM(dataKM, ci = TRUE, title = NULL)
}
\arguments{
\item{dataKM}{A dataframe containing the weigthed Kaplan-Meier estimates as computed by \code{\link[RiskStratifiedEstimation]{weightedKM}}}

\item{ci}{Should confidence intervals be displayed?}

\item{title}{The title on the graph}
}
\value{
A weighted Kaplan-Meier plot
}
\description{
Plots the weighted Kaplan-Meier estimate
}
