% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskReduction.R
\name{relativeRiskReduction}
\alias{relativeRiskReduction}
\title{Creates a data frame with hazard ratios within risk strata}
\usage{
relativeRiskReduction(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, truncatedWeights = TRUE,
  truncationQuantiles = c(0.01, 0.99))
}
\arguments{
\item{ps}{A list with the propensity data frames as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of weights to be used. Possible values are 'ATE' or 'ATT'}

\item{useStabilizedWeights}{Whether to use stabilized weights.}

\item{truncatedWeights}{Whether to truncate the weights}

\item{truncationQuantiles}{The quantiles used to truncate the weights}
}
\value{
A data frame with hazard ratios along with confidence intervals
}
\description{
Creates a data frame with hazard ratios within risk strata
}
