% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{relativeRiskReduction}
\alias{relativeRiskReduction}
\title{Relative risk reduction}
\usage{
relativeRiskReduction(model)
}
\arguments{
\item{model}{The model that was used to fit a cox regression model to the data.}
}
\value{
A dataframe with hazard ratios for treatment effect across risk strata along with 95 percent
                confidence intervals
}
\description{
Calculates hazard ratios within risk strata.
}
