% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskReduction.R
\name{relativeRiskReduction}
\alias{relativeRiskReduction}
\title{Creates a data frame with hazard ratios within risk strata}
\usage{
relativeRiskReduction(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, extremeWeights,
  fixedTruncationLevels = c(0.01, 0.99), truncationLevels = 0.1,
  cvLikeRepetitions, stepTruncationLevels)
}
\arguments{
\item{ps}{A list with the propensity data frames as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of the weights to be used. Allowed options are 'ATE' for average treatment effect and 'ATT' for average treatment effect on the treated weights}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{extremeWeights}{The way to assess extreme weights. Possible options are 'unadjusted, 'cvLikeTruncation', 'crumpTrimming', 'fixedTruncaiton'}

\item{fixedTruncationLevels}{The levels for fixed truncation weighting}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score. Only symmetric truncation is available. E.g. truncationLevels =.01 will assess truncation up to the .99th percentile of ps}

\item{cvLikeRepetitions}{The number of times to repeat the 2-fold cross-validations}

\item{stepTruncationLevels}{The steps for the grid of possible truncation levels}
}
\value{
A data frame with hazard ratios along with confidence intervals
}
\description{
Creates a data frame with hazard ratios within risk strata
}
