% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRiskStratifiedEstimation.R
\name{runRiskStratifiedEstimation}
\alias{runRiskStratifiedEstimation}
\title{Runs a risk stratified analysis}
\usage{
runRiskStratifiedEstimation(cohortMethodData, population, modelSettings, save,
  testSplit = "person", testFraction = 0.3, nfold = 10, riskStrata = 4,
  weightsType = "ATE", truncatedWeights = TRUE,
  useStabilizedWeights = FALSE, truncationQuantiles = c(0.01, 0.99),
  timePoint, excludeCovariateIds = NULL, binary = TRUE,
  includeAllOutcomes = TRUE, requireTimeAtRisk = TRUE, plpPlot = TRUE,
  psThreads = 1, priorType = "laplace", verbosity = "INFO",
  analysisId = NULL)
}
\arguments{
\item{cohortMethodData}{A cohortMethodData object}

\item{population}{The study population to perform the analysis}

\item{modelSettings}{The model settings for the prediction step}

\item{save}{The save directory}

\item{testSplit}{The type of split for the cross validation. Should be either 'person' or 'time'}

\item{testFraction}{The size of the test set}

\item{nfold}{The number of folds for cross validation}

\item{riskStrata}{The number of risk strata on which to perform the analysis}

\item{weightsType}{The type of weights for the balancing of covariates. Should be either 'ATE' or 'ATT'}

\item{truncatedWeights}{Should truncated weights be used?}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationQuantiles}{Quantiles of the propensity score to truncate in order to avoid excessively large weights}

\item{timePoint}{The time point of interest for the calculation of the absolute risk reduction}

\item{excludeCovariateIds}{Covariate Ids to be excluded from calculation of propensity scores}

\item{binary}{Forces the outcomeCount to be 0 or 1 in the prediction step}

\item{includeAllOutcomes}{(binary) indicating whether to include people with outcomes who are not observed for the whole at risk period}

\item{requireTimeAtRisk}{Should subjects without time at risk be removed at the prediction step?}

\item{plpPlot}{(binary) Should plots for the prediction step be generated?}

\item{psThreads}{The number of cores to use for the estimation of the propensity score. If 1 then serial approach is implemented}
}
\value{
\item{ps}{The propensity scores within risk strata along with patient weights}
\item{mapMatrix}{The matrix that maps the patients to risk strata}
\item{dataKM}{The weighted Kaplan-Meier estimates within risk strata}
\item{absoluteRiskReduction}{The absolute risk reduction within risk strata}
\item{relativeRiskReduction}{The relative risk reduction within risk strata}
}
\description{
Runs a risk stratified analysis
}
