% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRiskStratifiedEstimation.R
\name{runRiskStratifiedEstimation}
\alias{runRiskStratifiedEstimation}
\title{Runs a risk stratified analysis}
\usage{
runRiskStratifiedEstimation(cdmDatabaseSchema, cohortDatabaseSchema,
  outcomeDatabaseSchema, resultsDatabaseSchema, cohortTable, outcomeTable,
  mergedCohortTable, attributeDefinitionTable, cohortAttributeTable,
  treatmentCohortId, comparatorCohortId, outcomeIds, targetCohortId,
  connectionDetails, getDbCohortMethodDataArgs, covariateSettingsCm,
  populationCmSettings, exposureTable, psControl = NULL,
  psPrior = NULL, exposureDatabaseSchema, getPlpDataArgs,
  covariateSettingsPlp, modelSettings, populationPlpSettings,
  cdmVersion = "5", runPlpArgs, riskStrata = 4, weightsType = "ATE",
  useStabilizedWeights = TRUE, truncationLevels = c(0.01, 0.99),
  timePoint, compareAllOutcomes = TRUE, psThreads = 1,
  predictionThreads = 1, saveResults, saveDirectory = NULL, fftempdir,
  saveMapMatrix = TRUE, savePs = TRUE, verbosity = "INFO",
  analysisId = NULL)
}
\arguments{
\item{cdmDatabaseSchema}{The name of the database schema that contains the vocabulary files.
Requires read permissions to this database. On SQL Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that contains the treatment and comaparator cohorts. Requires read permissions
to this database.}

\item{outcomeDatabaseSchema}{The name of the database schema that contains the outcome cohorts. Requires read permissions to this database.}

\item{resultsDatabaseSchema}{The name of the database schema with write permissions.}

\item{cohortTable}{The name of the table holding the treatment and comparator cohorts.}

\item{outcomeTable}{The name of the table hodling the outcome cohorts.}

\item{mergedCohortTable}{The name of the table where the merged treatment and comparator cohorts will be stored.}

\item{attributeDefinitionTable}{The table where the definition of the treatment covariate will be stored.}

\item{cohortAttributeTable}{The table where the covariate values with regard to treatment will be stored.}

\item{treatmentCohortId}{The cohort definition id of the treatment cohort in the cohortTable.}

\item{comparatorCohortId}{The cohort definition id of the comparator cohort in the cohortTable.}

\item{outcomeIds}{A list of cohort definition ids used to define the outcomes in the outcome table.}

\item{targetCohortId}{The cohrt definition id of of the merged cohort in the mergedCohortTable.}

\item{connectionDetails}{An R object of type \code{connectionDetails} created using function
\code{\link[DatabaseConnector]{createConnectionDetails}} in the \code{DatabaseConnector} package.
Either the \code{connection} or the \code{connectionDetails} argument should be specified.}

\item{getDbCohortMethodDataArgs}{A parameter object for the function \code{\link[CohortMethod]{getDbCohortMethodData}}. Can be generated from function
\code{\link[CohortMethod]{createGetDbCohortMethodDataArgs}}.}

\item{covariateSettingsCm}{An object of type \code{covariateSettings} as created using the
\code{\link[FeatureExtraction]{createCovariateSettings}} to be used for the definition of the
\code{cohortMethodData} object.}

\item{populationCmSettings}{A parameter object for the function \code{\link[CohortMethod]{createStudyPopulation}}. Can be generated from
function \code{createStudyPopulationCmSettings}.}

\item{exposureTable}{Input of function \code{\link[CohortMethod]{getDbCohortMethodData}}: The tablename that contains the exposure cohorts.}

\item{psControl}{An object of the type \code{cyclopsControl} generated from \code{\link[Cyclops]{createControl}}.}

\item{psPrior}{An object of the type \code{cyclopsPrior} generated from \code{\link[Cyclops]{createPrior}}.}

\item{exposureDatabaseSchema}{Input of function \code{\link[CohortMethod]{getDbCohortMethodData}}: The name of the database schema that is the
location where the exposure data used to define the exposure cohorts is available.}

\item{getPlpDataArgs}{A parameter object for the function \code{\link[PatientLevelPrediction]{getPlpData}}. It can be generated from function
\code{\link[RiskStratifiedEstimation]{createGetPlpDataArgs}}.}

\item{covariateSettingsPlp}{An object of type \code{covariateSettings} as created using the
\code{\link[FeatureExtraction]{createCovariateSettings}} to be used for the definition of the
\code{plpData} object. note that a covariate indicating treatment will be added.}

\item{modelSettings}{An object of the class \code{modelSettings} to be used as input for \code{\link[PatientLevelPrediction]{runPlp}}.}

\item{populationPlpSettings}{A parameter object for the function \code{\link[PatientLevelPrediction]{createStudyPopulation}}. Can be generated from
function \code{\link[PatientLevelPrediction]{createStudyPopulationSettings}}.}

\item{cdmVersion}{Define the OMOP CDM version used: currently supported is "5".}

\item{runPlpArgs}{A parameter object for the function \code{\link[PatientLevelPrediction]{runPlp}}. Can be generated from function
\code{\link[RiskStratifiedEstimation]{createRunPlpArgs}}.}

\item{riskStrata}{The number of risk strata to divide the study population.}

\item{weightsType}{The type of weights for the balancing of covariates. Should be either 'ATE' or 'ATT'}

\item{useStabilizedWeights}{Should stabilized weights be used?}

\item{truncationLevels}{The level of truncation expressed in percentiles of the propensity score.}

\item{timePoint}{The time point of interest for the calculation of the absolute risk reduction.}

\item{compareAllOutcomes}{Should all the outcomes be analyzed within all stratifications?}

\item{psThreads}{The number of threads to be used for the estimation of the propensity scores.}

\item{predictionThreads}{The number of threads to be used to run the predictions.}

\item{saveResults}{Should the results of the entire analysis be saved?}

\item{saveDirectory}{The file path to the directory where the results of the analysis will be saved.}

\item{fftempdir}{The directory where the temporary \code{ff} files will be saved.}

\item{saveMapMatrix}{Should the map matrix with the risk sratum allocations be saved?}

\item{savePs}{Should the propensity scores be saved?}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold,
a message will print. The levels are:
\itemize{
    \item{DEBUG}{Highest verbosity showing all debug statements}
    \item{TRACE}{Showing information about start and end of steps}
    \item{INFO}{Show informative information (Default)}
    \item{WARN}{Show warning messages}
    \item{ERROR}{Show error messages}
    \item{FATAL}{Be silent except for fatal errors}}.}

\item{analysisId}{The identifier of the analysis.}
}
\value{
An object containing two large lists:
                                          \itemize{
                                              \item The propensity scores within risk strata for each outcome in \code{outcomeIds}
                                              \item The results of the risk stratified analyis within risk strata for all outcomes in \code{outcomeIds}:
                                                  \itemize{
                                                       \item Weighted Kaplan-Meier estimates.
                                                       \item Relative risk reduction.
                                                       \item Absolute risk reduction.
                                                       \item Number of cases.
                                                  }
                                           }
}
\description{
Runs a risk stratified analysis in two stages. It first runs a prediction algorithm using \code{PatientLevelPrediction} to derive baseline
patient risks and then derives estimates within risk strata using \code{CohortMethod} package.
}
