% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{runShiny}
\alias{runShiny}
\title{Launches the shiny application}
\usage{
runShiny(analysisSettings)
}
\arguments{
\item{analysisSettings}{An R object of type \code{analysisSettings} created using the function
\code{\link[RiskStratifiedEstimation]{createAnalysisSettings}}.}
}
\value{
The shiny application
}
\description{
Launches the shin application for the assessment of the overall results
}
