% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{weightedKM}
\alias{weightedKM}
\title{Calculates and plots weighted Kaplan-Meier estimates}
\usage{
weightedKM(ps, calculateWeights = TRUE, weightsType = "ATE",
  useStabilizedWeights = TRUE, truncatedWeights = TRUE,
  truncationQuantiles = c(0.01, 0.99), treatmentLabel = NULL,
  comparatorLabel = NULL)
}
\arguments{
\item{ps}{A data frame including the propensity scores as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of weights to be used. Possible values are 'ATE' or 'ATT'}

\item{useStabilizedWeights}{Whether to use stabilized weights.}

\item{truncatedWeights}{Whether to truncate the weights}

\item{truncationQuantiles}{The quantiles used to truncate the weights}

\item{treatmentLabel}{The label for the treatment cohort}

\item{comparatorLabel}{The label for the comparator cohort}
}
\value{
A data frame with the Kaplan-Meier estimates
}
\description{
Calculates the weighted Kaplan-Meier estimates based on: Xie J, Liu C. Adjusted Kaplan-Meier estimator and log-rank test with inverse probability of treatment weighting for survival data. Statistics in Medicine 2005; 2:3089–3110.
}
