% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{weightedKM}
\alias{weightedKM}
\title{Calculates and plots weighted Kaplan-Meier estimates}
\usage{
weightedKM(ps, calculateWeights = TRUE, weightsType = "ATE", useSW = TRUE,
  truncatedWeights = TRUE, truncationQuantiles = c(0.01, 0.99),
  returnPlot = TRUE, xlim = NULL, ylim = NULL, title = NULL,
  legend.position = c(0.15, 0.1), ci = FALSE)
}
\arguments{
\item{ps}{A data frame including the propensity scores as generated from \code{\link[CohortMethod]{createPs}}}

\item{calculateWeights}{Whether to calculate the weights using \code{\link[RiskStratifiedEstimation]{createIPW}}}

\item{weightsType}{The type of weights to be used. Possible values are 'ATE' or 'ATT'}

\item{useSW}{Whether to use stabilized weights.}

\item{truncatedWeights}{Whether to truncate the weights}

\item{truncationQuantiles}{The quantiles used to truncate the weights}

\item{returnPlot}{Whether to return Kaplan-Meier plot}

\item{xlim}{Limits on x-axis}

\item{ylim}{Limits on y-axis}

\item{title}{Title appearing in the plot}

\item{legend.position}{The position of the legend}

\item{ci}{Whether the confidence intervals should appear in  the plot}
}
\value{
A data frame with the Kaplan-Meier estimates
}
\description{
Calculates the weighted Kaplan-Meier estimates based on: Xie J, Liu C. Adjusted Kaplan-Meier estimator and log-rank test with inverse probability of treatment weighting for survival data. Statistics in Medicine 2005; 2:3089–3110.
}
